Separate authentication and user services
